Bug fixes and OSX Build Updates
Lots of fixes here, reworked OSX and fixed several issues. Create
Makefile and updates README to include OSX build instructions. Also
created a new function to reset the texture filtering for
FULLSCREEN_DESKTOP mode.  This has to be used in OSX/Linux due to issue
with normal FULLSCREEN.   FULLSCREEN_DESKTOP doesn't scale so texture
filter has to be used to keep the proper aspect ratio for shaded blocks,
otherwise it's a mess.  This causes a slight blur until i can code or
work in a sampling algo.