ScreenBuffer functions to mimic screen drawing.
ScreenBuffer functions to mimic screen drawing., for copy/paste to work
correctly the screen buffer much match the screen, so on character
erases and scrolling the same must be done to the buffer, this keeps it
quick if were not drawing the entire buffer each character change.
Still need to add functions for scrolling region.