Mouse Selection Fix for clicks on the Title Bar
Switching focus out of the app then back to the app by clicking on the
title bar will no longer keep the mouse registered as text selection.
This is a work around since the OS seems to send a MOUSEDOWN event
without a proper MOUSEUP even on the release.   The mouse registers the
y position as 0, so we avoid setting the trigger if this is the
coordinate sent from the mouse click event.