Initial FreeBSD port.
Fix up for more standard make arguments, and allow overriding them.
This allows corss-compiling and building on systems that don't have
a C++ compiler installed as /usr/bin/g++.

The method of finding the executable path is very fragile on Linux
(and now FreeBSD) and it should instead parse argv[0] and getcwd().