Change Log 3.0.22 - 2016-01-05
- Created new Fontset.xml file to hold all system fonts with names,
height,
width, filename and ESC sequence for automatic font switching.
The availability to set the sequence in the XML is an added bonus to set
a preference.  For example.  If ESC[0;40 D is received for Amiga Topaz,
however you preference is to see Topaz Plus, you can set this ESC
sequence
under the preferred font so it will be used instead.  All fonts are now
configurable.
- Also updated the Dialing Directory to pull front names from the new
xml set.