Change Log 3.0.29 - 2016-01-17 / 18
- Starting on Connections
- Adding Creating and Connection Instance to SessionManager Queue from
MenuManager.
- Connections are working, however need to add in telnet option passing
to handle
incoming data properly.  Also have to make sure Async Sockets in
separate thread
are allowing the main thread to process the data properly.
- Added In Telnet Definitions and created new TelnetManager class for
option negotiations.
- Also added new TermType to the DialDirectoy XML file.  This is the
Terminal type sent to the
server on connections ANSI is default, others could be XTERM depend on
the shell or server
* ISSUE:
- Just realized that Connection Swaps are done in the async.connection
callback
which is threaded with IO_Service.  SDL Events can't hit Windows that
are swapped in threads
go figure.  So right now it connects and renders, but then freezes
because it can't handle
events.  So this needs to be reworked to register a connection, passing
the socket, then
a new session needs to be swapped separately from the Interface class
not in a thread! WIP!